jac-module-proxy/pyproject.toml
[tool.poetry]
name = "bcsai-be"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "server"}]


[tool.poetry.dependencies]
python = "3.13.2"
grpcio = "1.71.0"
grpcio-tools = "1.71.0"
grpcio-health-checking = "1.71.0"
uuid6 = "2024.7.10"

[tool.poetry.group.dev.dependencies]
python-dotenv = "1.0.1"
black = "25.1.0"
pre-commit = "4.1.0"
pytest = "8.3.5"
flake8 = "7.1.2"
flake8-import-order = "0.18.2"
flake8-docstrings = "1.7.0"
flake8-comprehensions = "3.16.0"
flake8-bugbear = "24.12.12"
flake8-annotations = "3.1.1"
flake8_simplify = "0.21.0"
pep8-naming = "0.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
deploy = "server.main:run"